requirements.txt
flask
flask-cors
pandas
sentence-transformers
chromadb
uuid
sqlite3
requests
websocket-client
